Polish ActivitiGraphQLQueryService Spring Boot Auto-configuration
